+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod event_id;
 pub mod model;
 pub mod raw;
 pub mod traits;